1.0.3
Tweaked how critters spawn in front of the caster to reduce the chance of physical collisions with freshly spawned critters while moving

Added an MCM setting to further customize the above; critters can be made to spawn closer, further away, or even behind the caster if desired

Added a new MCM option, "Critter Return Distance Offset"; this controls how closely to the caster that critters will return before going "poof", to reduce the chances of the critters knocking the player around.

The default setting for the above is fairly conservative (150 game units), but this can be adjusted higher or lower depending on how annoying you find getting bumped around to be.

Increased the volume of the despawn "poof" sound effect slightly

A few minor script optimizations here and there

Vanilla Fixes module: corrected an error in one of my leveled lists which caused planted Blisterwort to produce Namira's Rot instead